backend/requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
chromadb
tavily-python
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
langchain-google-genai
pypdf
fastapi
uvicorn
alembic
psycopg2-binary
apscheduler